% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaodData.R
\name{BreastCancerATAC}
\alias{BreastCancerATAC}
\title{Laod TCGA Breast Cancer ATAC-seq expression matrix.}
\usage{
BreastCancerATAC()
}
\value{
ATAC-seq count
}
\description{
Laod TCGA Breast Cancer ATAC-seq expression matrix.
}
\examples{
library(LinkageData)
BreastCancerATAC()
}
