% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaodData.R
\name{BreastCancerRNA}
\alias{BreastCancerRNA}
\title{Laod TCGA Breast Cancer RNA-seq expression matrix.}
\usage{
BreastCancerRNA()
}
\value{
RNA-seq count
}
\description{
Laod TCGA Breast Cancer RNA-seq expression matrix.
}
\examples{
library(LinkageData)
BreastCancerRNA()
}
