% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaodData.R
\name{MuSCsRNA}
\alias{MuSCsRNA}
\title{Load MuSC RNA-seq expression matrix.}
\usage{
MuSCsRNA()
}
\value{
RNA-seq count
}
\description{
Load MuSC RNA-seq expression matrix.
}
\examples{
library(LinkageData)
MuSCsRNA()
}
