% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaodData.R
\name{MuSCsATAC}
\alias{MuSCsATAC}
\title{Load MuSC ATAC-seq expression matrix.}
\usage{
MuSCsATAC()
}
\value{
ATAC-seq count
}
\description{
Load MuSC ATAC-seq expression matrix.
}
\examples{
library(LinkageData)
MuSCsATAC()
}
